pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trustquerynet"
version = "0.1.0"
description = "Local-first trustworthy image classification pipeline for noisy labels, uncertainty, active querying, and selective prediction."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
  "timm>=0.9.12",
  "torch>=2.2",
  "torchvision>=0.17",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["trustquerynet*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
